update rm and add cat

+++ src/Terminal.java
@@ -38,6 +38,8 @@ public class Terminal {
             echo(commandArgs);
         } else if (commandName.equals("rm")) {
             rm(commandArgs);
+        } else if (commandName.equals("cat")) {
+            cat(commandArgs);
         } else if (commandName.equals("exit")) {
             System.exit(0);
         }
@@ -53,14 +55,49 @@ public class Terminal {
     }
 
     public void rm(String[] args) {
-        String file = args[0];
-        try {
-            Path filePath = currentDirectory.resolve(file);
-            Files.delete(filePath);
-        } catch (NoSuchFileException e) {
-            System.out.println("rm: cannot remove '" + file + "': No such file or directory");
-        } catch (IOException e) {
-            System.out.println("rm: cannot remove '" + file + "': Permission denied");
+        if(args.length == 1){
+            String file = args[0];
+            try {
+                Path filePath = currentDirectory.resolve(file);
+                if (Files.isRegularFile(filePath)) {
+                    Files.delete(filePath);
+                } else {
+                    System.out.println("rm: cannot remove '" + file + "': Not a file");
+                }
+            } catch (NoSuchFileException e) {
+                System.out.println("rm: cannot remove '" + file + "': No such file or directory");
+            } catch (IOException e) {
+                System.out.println("rm: cannot remove '" + file + "': Permission denied");
+            }
+        }else{
+            System.out.println("rm: Invalid number of arguments");
+        }
+    }
+
+    public void cat(String[] args){
+        if (args.length == 1) {
+            try {
+                Path filePath = currentDirectory.resolve(args[0]);
+                System.out.println(Files.readString(filePath));
+            } catch (NoSuchFileException e) {
+                System.out.println("cat: " + args[0] + ": No such file or directory");
+            } catch (IOException e) {
+                System.out.println("cat: " + args[0] + ": Error reading the file");
+            }
+        } else if (args.length == 2) {
+            try {
+                Path filePath1 = currentDirectory.resolve(args[0]);
+                Path filePath2 = currentDirectory.resolve(args[1]);
+                String content1 = Files.readString(filePath1);
+                String content2 = Files.readString(filePath2);
+                System.out.println(content1 + '\n' + content2);
+            } catch (NoSuchFileException e) {
+                System.out.println("cat: No such file or directory");
+            } catch (IOException e) {
+                System.out.println("cat: Error reading the file");
+            }
+        } else {
+            System.out.println("cat: Invalid number of arguments");
         }
     }
